+++ HGF/tapas_softmax_mu3.m
@@ -2,7 +2,7 @@ function [logp, yhat, res] = tapas_softmax_mu3(r, infStates, ptrans)
 % Calculates the log-probability of responses under the softmax model
 %
 % --------------------------------------------------------------------------------------------------
-% Copyright (C) 2017 Christoph Mathys, TNU, UZH & ETHZ
+% Copyright (C) 2017-2019 Christoph Mathys, TNU, UZH & ETHZ
 %
 % This file is part of the HGF toolbox, which is released under the terms of the GNU General Public
 % Licence (GPL), version 3. You can redistribute it and/or modify it under the terms of the GPL
@@ -29,18 +29,24 @@ res  = NaN(n,1);
 % dim 3: choice number
 % dim 4: 1: muhat, 2: sahat, 3: mu, 4: sa
 
-% Weed irregular trials out from inferred states and responses
-states = squeeze(infStates(:,1,:,pop));
-states(r.irr,:) = [];
-mu3 = squeeze(infStates(:,3,1,3));
-mu3(r.irr) = [];
-y = r.y(:,1);
-y(r.irr) = [];
-
 % Number of choices
 nc = size(infStates,3);
 
-% Inveyrse decision temperature
+% Belief trajectories at 1st level
+states = squeeze(infStates(:,1,:,pop));
+
+% Log-volatility trajectory
+mu3 = squeeze(infStates(:,3,1,3));
+
+% Responses
+y = r.y(:,1);
+
+% Weed irregular trials out from inferred states and responses
+states(r.irr,:) = [];
+mu3(r.irr) = [];
+y(r.irr) = [];
+
+% Inverse decision temperature
 be = exp(-mu3);
 be = repmat(be,1,nc);
 
@@ -60,4 +66,4 @@ logp(reg) = log(probc);
 yhat(reg) = probc;
 res(reg) = -log(probc);
 
-return;
+end