+++ PhysIO/examples/tapas_physio_example_siemens_vd_ppu3t_spm_job.m
@@ -1,0 +1,41 @@
+%-----------------------------------------------------------------------
+% Job saved on 05-Sep-2018 03:20:54 by cfg_util (rev $Rev: 6942 $)
+% spm SPM - SPM12 (7219)
+% cfg_basicio BasicIO - Unknown
+%-----------------------------------------------------------------------
+matlabbatch{1}.spm.tools.physio.save_dir = {''};
+matlabbatch{1}.spm.tools.physio.log_files.vendor = 'Siemens_Tics';
+matlabbatch{1}.spm.tools.physio.log_files.cardiac = {'Physio_20180101_120001_pppppppp-hhhh-yyyy-ssss-ioioioioioio_PULS.log'};
+matlabbatch{1}.spm.tools.physio.log_files.respiration = {'Physio_20180101_120001_pppppppp-hhhh-yyyy-ssss-ioioioioioio_RESP.log'};
+matlabbatch{1}.spm.tools.physio.log_files.scan_timing = {'Physio_20180101_120001_pppppppp-hhhh-yyyy-ssss-ioioioioioio_Info.log'};
+matlabbatch{1}.spm.tools.physio.log_files.sampling_interval = [];
+matlabbatch{1}.spm.tools.physio.log_files.relative_start_acquisition = 0;
+matlabbatch{1}.spm.tools.physio.log_files.align_scan = 'last';
+matlabbatch{1}.spm.tools.physio.scan_timing.sqpar.Nslices = 48;
+matlabbatch{1}.spm.tools.physio.scan_timing.sqpar.NslicesPerBeat = [];
+matlabbatch{1}.spm.tools.physio.scan_timing.sqpar.TR = 0.995;
+matlabbatch{1}.spm.tools.physio.scan_timing.sqpar.Ndummies = 0;
+matlabbatch{1}.spm.tools.physio.scan_timing.sqpar.Nscans = 30;
+matlabbatch{1}.spm.tools.physio.scan_timing.sqpar.onset_slice = 1;
+matlabbatch{1}.spm.tools.physio.scan_timing.sqpar.time_slice_to_slice = [];
+matlabbatch{1}.spm.tools.physio.scan_timing.sqpar.Nprep = [];
+matlabbatch{1}.spm.tools.physio.scan_timing.sync.scan_timing_log = struct([]);
+matlabbatch{1}.spm.tools.physio.preproc.cardiac.modality = 'PPU';
+matlabbatch{1}.spm.tools.physio.preproc.cardiac.initial_cpulse_select.auto_matched.min = 0.4;
+matlabbatch{1}.spm.tools.physio.preproc.cardiac.initial_cpulse_select.auto_matched.file = 'initial_cpulse_kRpeakfile.mat';
+matlabbatch{1}.spm.tools.physio.preproc.cardiac.posthoc_cpulse_select.off = struct([]);
+matlabbatch{1}.spm.tools.physio.model.output_multiple_regressors = 'multiple_regressors.txt';
+matlabbatch{1}.spm.tools.physio.model.output_physio = 'physio.mat';
+matlabbatch{1}.spm.tools.physio.model.orthogonalise = 'none';
+matlabbatch{1}.spm.tools.physio.model.censor_unreliable_recording_intervals = false;
+matlabbatch{1}.spm.tools.physio.model.retroicor.yes.order.c = 3;
+matlabbatch{1}.spm.tools.physio.model.retroicor.yes.order.r = 4;
+matlabbatch{1}.spm.tools.physio.model.retroicor.yes.order.cr = 1;
+matlabbatch{1}.spm.tools.physio.model.rvt.no = struct([]);
+matlabbatch{1}.spm.tools.physio.model.hrv.no = struct([]);
+matlabbatch{1}.spm.tools.physio.model.noise_rois.no = struct([]);
+matlabbatch{1}.spm.tools.physio.model.movement.no = struct([]);
+matlabbatch{1}.spm.tools.physio.model.other.no = struct([]);
+matlabbatch{1}.spm.tools.physio.verbose.level = 2;
+matlabbatch{1}.spm.tools.physio.verbose.fig_output_file = '';
+matlabbatch{1}.spm.tools.physio.verbose.use_tabs = false;